Fix the daymail image bug.
Since the asset() function is adding "?v*.*.*" at the end of the resource link, the daymail script can't fetch the picture while parsing the body to find them. It occurs when the user copy/paste the image URL to insert it into a daymail.

Instead of patching the DaymailModule, I prefer to correct the display, since the version is not really useful in emails.